+++ packages/core/src/defaults/values/index.ts
@@ -1,5 +1,6 @@
 export * from "./border-values";
 export * from "./box-model-values";
 export * from "./color-values";
+export * from "./effect-values";
 export * from "./grid-values";
 export * from "./interactivity-values";